Reorganize IMU and quaternion math functions
