+++ components/hdw-imu/include/hdw-imu.h
@@ -46,6 +46,8 @@
 #include <hal/gpio_types.h>
 #include <esp_err.h>
 
+#include "quaternions.h"
+
 typedef struct
 {
     int32_t temp;
@@ -88,20 +90,5 @@ esp_err_t accelIntegrate(void);
 esp_err_t accelPerformCal(void);
 float accelGetStdDevInCal(void);
 void accelSetRegistersAndReset(void);
-
-// Utility functions (to replace at a later time)
-
-float rsqrtf(float x);
-float mathsqrtf(float x);
-void mathEulerToQuat(float* q, const float* euler);
-void mathQuatApply(float* qout, const float* q1, const float* q2);
-void mathQuatNormalize(float* qout, const float* qin);
-void mathCrossProduct(float* p, const float* a, const float* b);
-void mathRotateVectorByInverseOfQuaternion(float* pout, const float* q, const float* p);
-void mathRotateVectorByQuaternion(float* pout, const float* q, const float* p);
-esp_err_t GeneralSet(int dev, int reg, int val);
-esp_err_t LSM6DSLSet(int reg, int val);
-int GeneralI2CGet(int device, int reg, uint8_t* data, int data_len);
-int ReadLSM6DSL(uint8_t* data, int data_len);
 
 #endif
